+++ modules/sagemaker/sagemaker-templates/stack.py
@@ -50,6 +50,7 @@ class ProjectStack(Stack):
         hf_import_models_project_settings: Any = None,
         batch_inference_project_settings: Any = None,
         permissions_boundary_name: Optional[str] = None,
+        s3_access_logs_bucket_arn: Optional[str] = None,
         **kwargs: Any,
     ) -> None:
         super().__init__(scope, id, **kwargs)
@@ -87,6 +88,7 @@ class ProjectStack(Stack):
                 access_token_secret_name=access_token_secret_name,
                 aws_codeconnection_arn=aws_codeconnection_arn,
                 repository_owner=repository_owner,
+                s3_access_logs_bucket_arn=s3_access_logs_bucket_arn,
             )
         elif project_template_type == ProjectTemplateType.BATCH_INFERENCE:
             BatchInferenceProject(
@@ -104,6 +106,7 @@ class ProjectStack(Stack):
                 access_token_secret_name=access_token_secret_name,
                 aws_codeconnection_arn=aws_codeconnection_arn,
                 repository_owner=repository_owner,
+                s3_access_logs_bucket_arn=s3_access_logs_bucket_arn,
             )
         elif project_template_type == ProjectTemplateType.FINETUNE_LLM_EVALUATION:
             FinetuneLlmEvaluationProject(